model/requirements.txt
pandas
scikit-learn
joblib
azure-identity
azure-storage-blob
